refactor: changed useScrollToChange hook file name
